Moved contract ConsensusCogateway to consensus-gateway folder;
