+++ test/consensus/utils.js
@@ -15,7 +15,6 @@
 'use strict';
 
 const web3 = require('../test_lib/web3.js');
-const Utils = require('../test_lib/utils.js');
 
 const SentinelCommittee = '0x0000000000000000000000000000000000000001';
 const CommitteeFormationDelay = 14;
@@ -116,44 +115,6 @@ function getMetachainIdHash(anchor, metachainIdTypeHash) {
       [metachainIdTypeHash, anchor],
     ),
   );
-}
-
-function getMessageInboxIdentifier(metachainId, verifyingAddress) {
-  return web3.utils.sha3(Utils.encodeParameters(
-    [
-      'bytes32',
-      'string',
-      'string',
-      'bytes32',
-      'address',
-    ],
-    [
-      DOMAIN_SEPARATOR_TYPEHASH,
-      MESSAGE_BUS_DOMAIN_SEPARATOR_NAME,
-      DOMAIN_SEPARATOR_VERSION,
-      metachainId,
-      verifyingAddress,
-    ],
-  ));
-}
-
-function getMessageOutboxIdentifier(metachainId, verifyingAddress) {
-  return web3.utils.sha3(Utils.encodeParameters(
-    [
-      'bytes32',
-      'string',
-      'string',
-      'bytes32',
-      'address',
-    ],
-    [
-      DOMAIN_SEPARATOR_TYPEHASH,
-      MESSAGE_BUS_DOMAIN_SEPARATOR_NAME,
-      DOMAIN_SEPARATOR_VERSION,
-      metachainId,
-      verifyingAddress,
-    ],
-  ));
 }
 
 module.exports = {
